refactor(BotBlessed): rename botBlessed class to BotBlessed
